Виправнено помилки
Додано:
.gitignore
requirements.txt
клас для повідомлень

Видалено:
папку user

Змінено:
logging.INFO на logging.DEBUG
