+++ main.py
@@ -1,4 +1,4 @@
-from user import user
+import user
 from create_bot import bot, dp
 import asyncio
 import logging
@@ -16,5 +16,5 @@ async def main():
 
 if __name__ == "__main__":
     user.register(dp)
-    logging.basicConfig(level=logging.INFO, stream=sys.stdout)
+    logging.basicConfig(level=logging.DEBUG, stream=sys.stdout)
     asyncio.run(main())
