+++ message_text.py
@@ -1,0 +1,10 @@
+from aiogram.types import Message
+
+
+class User_message_text:
+
+    def hello_user_message(self) -> str:
+        return 'Привіт'
+
+    def eho_user_message(self, text) -> str:
+        return text
