--- user/message_text.py
@@ -1,1 +1,0 @@
-hello_message = "Привіт"
